+++ src/views.jl
@@ -1,30 +1,74 @@
-"""
-    BlockSlice(indices)
+### Views
+function _unblock(cum_sizes, I::Tuple{BlockIndexRange{1,R}, Vararg{Any}}) where {R}
+    B = Block(first(I))
+    range = cum_sizes[Int(B)]-1 .+ first(I).indices[1]
 
-Represent an AbstractUnitRange of indices that attaches a block.
-
-Upon calling `to_indices()`, Blocks are converted to BlockSlice objects to represent
-the indices over which the Block spans.
-
-This mimics the relationship between `Colon` and `Base.Slice`.
-"""
-struct BlockSlice{BB} <: AbstractUnitRange{Int}
-    block::BB
-    indices::UnitRange{Int}
+    BlockSlice(I[1], range)
 end
 
-Block(bs::BlockSlice{<:Block}) = bs.block
+
+to_index(::BlockIndexRange) = throw(ArgumentError("BlockIndexRange must be converted by to_indices(...)"))
+
+@inline to_indices(A, inds, I::Tuple{BlockIndexRange{1,R}, Vararg{Any}}) where R =
+    (unblock(A, inds, I), to_indices(A, _maybetail(inds), tail(I))...)
+
+# splat out higher dimensional blocks
+# this mimics view of a CartesianIndex
+@inline to_indices(A, inds, I::Tuple{BlockIndexRange, Vararg{Any}}) =
+    to_indices(A, inds, (BlockRange.(Block.(I[1].block.n), tuple.(I[1].indices))..., tail(I)...))
 
 
-for f in (:axes, :unsafe_indices, :axes1, :first, :last, :size, :length,
-          :unsafe_length, :start)
-    @eval $f(S::BlockSlice) = $f(S.indices)
+# In 0.7, we need to override to_indices to avoid calling linearindices
+@inline to_indices(A, I::Tuple{BlockIndexRange, Vararg{Any}}) =
+    to_indices(A, axes(A), I)
+
+if VERSION >= v"1.2-"  # See also `reindex` definitions in views.jl
+reindex(idxs::Tuple{BlockSlice{<:BlockRange}, Vararg{Any}},
+        subidxs::Tuple{BlockSlice{<:BlockIndexRange}, Vararg{Any}}) =
+    (@_propagate_inbounds_meta; (BlockSlice(BlockIndexRange(Block(idxs[1].block.indices[1][Int(subidxs[1].block.block)]),
+                                                            subidxs[1].block.indices),
+                                            idxs[1].indices[subidxs[1].indices]),
+                                 reindex(tail(idxs), tail(subidxs))...))
+else  # if VERSION >= v"1.2-"
+reindex(V, idxs::Tuple{BlockSlice{<:BlockRange}, Vararg{Any}},
+        subidxs::Tuple{BlockSlice{<:BlockIndexRange}, Vararg{Any}}) =
+    (@_propagate_inbounds_meta; (BlockSlice(BlockIndexRange(Block(idxs[1].block.indices[1][Int(subidxs[1].block.block)]),
+                                                            subidxs[1].block.indices),
+                                            idxs[1].indices[subidxs[1].indices]),
+                                    reindex(V, tail(idxs), tail(subidxs))...))
+end  # if VERSION >= v"1.2-"
+
+# De-reference blocks before creating a view to avoid taking `global2blockindex`
+# path in `AbstractBlockStyle` broadcasting.
+@inline function Base.unsafe_view(
+        A::BlockArray{<:Any, N},
+        I::Vararg{BlockSlice{<:BlockIndexRange{1}}, N}) where {N}
+    @_propagate_inbounds_meta
+    B = A[map(x -> x.block.block, I)...]
+    return view(B, _splatmap(x -> x.block.indices, I)...)
 end
 
-getindex(S::BlockSlice, i::Integer) = getindex(S.indices, i)
-show(io::IO, r::BlockSlice) = print(io, "BlockSlice(", r.block, ",", r.indices, ")")
-next(S::BlockSlice, s) = next(S.indices, s)
-done(S::BlockSlice, s) = done(S.indices, s)
+@inline function Base.unsafe_view(
+        A::PseudoBlockArray{<:Any, N},
+        I::Vararg{BlockSlice{<:BlockIndexRange{1}}, N}) where {N}
+    @_propagate_inbounds_meta
+    return view(A.blocks, map(x -> x.indices, I)...)
+end
+
+@inline function Base.unsafe_view(
+        A::ReshapedArray{<:Any, N, <:AbstractBlockArray{<:Any, M}},
+        I::Vararg{BlockSlice{<:BlockIndexRange{1}}, N}) where {N, M}
+    @_propagate_inbounds_meta
+    # Note: assuming that I[M+1:end] are verified to be singletons
+    return reshape(view(A.parent, I[1:M]...), Val(N))
+end
+
+@inline function Base.unsafe_view(
+        A::Array{<:Any, N},
+        I::Vararg{BlockSlice{<:BlockIndexRange{1}}, N}) where {N}
+    @_propagate_inbounds_meta
+    return view(A, map(x -> x.indices, I)...)
+end
 
 function _unblock(cum_sizes, I::Tuple{Block{1, T},Vararg{Any}}) where {T}
     B = first(I)
